pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "holoflow"
version = "0.1.0"
description = "Real-time Doppler holography"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
  "numpy>=2.4.4",
  "cupy-cuda13x>=14.0.1",
  "PySide6>=6.11.0",
  "PyOpenGL>=3.1.10",
  "pyyaml>=6.0",
  "holofile",
  "viztracer>=1.1.1",
  "matplotlib>=3.10.8",
  "tabulate>=0.10.0",
]

[project.scripts]
holoflow = "holoflow.main:main"
cupy_naive = "cupy_naive.main:main"
cupy_threaded = "cupy_threaded.main:main"
cupy_streams = "cupy_streams.main:main"

[tool.hatch.build.targets.wheel]
packages = [
  "src/holoflow",
  "src/holoflow_benchmarks",
  "src/cupy_naive",
  "src/cupy_threaded",
  "src/cupy_streams",
]

[tool.uv.sources]
holofile = { git = "https://github.com/JulesGuillou/holofile.git", rev = "main" }
